+++ Sources/Sentry/include/SentrySwizzle.h
@@ -350,6 +350,20 @@ typedef NS_ENUM(NSUInteger, SentrySwizzleMode) {
 // and remove it later.
 #define _SentrySWArguments(arguments...) DEL, ##arguments
 
+#if TEST
+#    define _SentrySWReplacement(code...)                                                          \
+        @try {                                                                                     \
+            code                                                                                   \
+        } @finally {                                                                               \
+            if (!swizzleInfo.originalCalled)                                                       \
+                @throw([NSException exceptionWithName:@"SwizzlingError"                            \
+                                               reason:@"Original method not called"                \
+                                             userInfo:nil]);                                       \
+        }
+#else
+#    define _SentrySWReplacement(code...) code
+#endif
+
 #define _SentrySwizzleInstanceMethod(classToSwizzle, selector, SentrySWReturnType,                 \
     SentrySWArguments, SentrySWReplacement, SentrySwizzleMode, KEY)                                \
     [SentrySwizzle                                                                                 \
@@ -359,18 +373,8 @@ typedef NS_ENUM(NSUInteger, SentrySwizzleMode) {
                     SentrySWReturnType (*originalImplementation_)(                                 \
                         _SentrySWDel3Arg(__unsafe_unretained id, SEL, SentrySWArguments));         \
                     SEL selector_ = selector;                                                      \
-                    return ^SentrySWReturnType(                                                    \
-                        _SentrySWDel2Arg(__unsafe_unretained id self, SentrySWArguments)) {        \
-                        @try {                                                                     \
-                            SentrySWReplacement                                                    \
-                        } @finally {                                                               \
-                            if (!swizzleInfo.originalCalled)                                       \
-                                @throw(                                                            \
-                                    [NSException exceptionWithName:@"SwizzlingError"               \
-                                                            reason:@"Original method not called"   \
-                                                          userInfo:nil]);                          \
-                        }                                                                          \
-                    };                                                                             \
+                    return ^SentrySWReturnType(_SentrySWDel2Arg(__unsafe_unretained id self,       \
+                        SentrySWArguments)) { _SentrySWReplacement(SentrySWReplacement) };         \
                 }                                                                                  \
                          mode:SentrySwizzleMode                                                    \
                           key:KEY];
@@ -384,17 +388,8 @@ typedef NS_ENUM(NSUInteger, SentrySwizzleMode) {
                  SentrySWReturnType (*originalImplementation_)(                                    \
                      _SentrySWDel3Arg(__unsafe_unretained id, SEL, SentrySWArguments));            \
                  SEL selector_ = selector;                                                         \
-                 return ^SentrySWReturnType(                                                       \
-                     _SentrySWDel2Arg(__unsafe_unretained id self, SentrySWArguments)) {           \
-                     @try {                                                                        \
-                         SentrySWReplacement                                                       \
-                     } @finally {                                                                  \
-                         if (!swizzleInfo.originalCalled)                                          \
-                             @throw([NSException exceptionWithName:@"Swizzling Error"              \
-                                                            reason:@"Original method not called"   \
-                                                          userInfo:nil]);                          \
-                     }                                                                             \
-                 };                                                                                \
+                 return ^SentrySWReturnType(_SentrySWDel2Arg(__unsafe_unretained id self,          \
+                     SentrySWArguments)) { _SentrySWReplacement(SentrySWReplacement) };            \
              }];
 
 #define _SentrySWCallOriginal(arguments...)                                                        \