+++ Tests/SentryTests/Integrations/SentryUIViewControllerSwizziling+Test.h
@@ -4,7 +4,8 @@ NS_ASSUME_NONNULL_BEGIN
 
 #if SENTRY_HAS_UIKIT
 
-@interface SentryUIViewControllerSwizziling (Test)
+@interface
+SentryUIViewControllerSwizziling (Test)
 
 - (BOOL)shouldSwizzleViewController:(Class)class;