+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 7.4.5
+
+- fix: Remove Check for Original Method Call When Swizzling (#1383)
+- fix: Init for Span, Tracer, Transaction (#1385)
+
 ## 7.4.4
 
 - fix: Crash for Call Should be on Main Thread (#1371)